archiver/requirements.txt
google-cloud-bigquery==3.13.0
google-cloud-storage==2.13.0
pandas==2.1.3
pyarrow==14.0.1